Rename
renamed all remnates of racadm to chassism1000e for the new class name
