+++ lib/puppet/util/network_device/chassism1000e.rb
@@ -1,0 +1,4 @@
+
+module Puppet::Util::NetworkDevice::Chassism1000e
+
+end
