--- lib/puppet/util/network_device/racadm.rb
@@ -1,4 +1,0 @@
-
-module Puppet::Util::NetworkDevice::Racadm
-
-end
