+++ templates/config.erb
@@ -1,3 +1,3 @@
 [<%= name %>]
-type racadm
+type chassism1000e
 url ssh://<%= username %>:<%= password %>@<%= url %>:<%= port %>
